backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.8.6
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==2.7.0
python-dotenv==1.0.1
pydantic==2.7.1
packaging==24.0
httpx==0.27.0
